Eigvals are floats (#5607)
**Context:**
Most eigvals are floats, but specifically for X, Y, Z and Hadamard, they
have been (1, -1). This can make trouble - for example, it caused an
issue with JAX-JIT, because it doesn't know what dtype to anticipate
coming from a sampled observable.

**Description of the Change:**
Main change here is changing eigvals to (1.0, -1.0) so that all
observables have the same dtype for eigenvalues.
There is still some special handling for X, Y, Z and Hadamard, because
it makes things JIT compatible for those operators, but it now pulls
from the `op.eigvals()` property and returns floats.

**Benefits:**

- Sampling of observables is more consistent in how it is processed and
what dtype is returned
- JAX-JIT with samples and an observable works now, see linked issue
below

**Possible Drawbacks:**
If people were doing checks for `int`, or otherwise assuming they would
get integers when returning fx. `qml.sample(qml.X(0))`, this will be a
breaking change for them.

**Related GitHub Issues:**
#5369

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Vincent Michaud-Rioux <vincentm@nanoacademic.com>
Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>