Importing the `default.tensor` in PL from the `lightning.tensor` device in PL lightning (#5699)
**Context:** This PR moves part of the `lightning.tensor` device from
PennyLane lightning to PennyLane. Such a device was introduced
originally in the following pull requests:
[1](https://github.com/PennyLaneAI/pennylane-lightning/pull/671) and
[2](https://github.com/PennyLaneAI/pennylane-lightning/pull/686). Due to
requirements change, it has been decided to create an additional
`default.tensor` device and move it to the PennyLane repository.
Specifically, we move the `quimb` interface and the corresponding Python
implementation of `lightning.tensor` into `default.tensor`. For
reference, the PR removing such a component from `lightning.tensor` is
[here](https://github.com/PennyLaneAI/pennylane-lightning/pull/737).

*Please note that the full documentation for such a device will be added
in a following PR*

**Description of the Change:** As above.

**Benefits:** Having two separate devices simplifies the distinction
between the C++ and the Python components. The relationship between
`lightning.tensor` and `default.tensor` will be similar to that between
`lightning.qubit` and `default.qubit`.

**Possible Drawbacks:** None with the existing code, since this is a new
quantum device.

**Related GitHub Issues:** None.

**Related Shortcut Stories:**
[sc-63256]
[sc-62979]

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>