Add prelim support for test-level markers (#5517)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** This PR introduces new test-level markers for PennyLane,
which can begin use for indication of where interactions take-place in
the given suite. This forms part of the work on improving testing across
the packages
[sc-61416]

**Description of the Change:** Add 3 new test-level markers for pytest:
`unit_test`, `integration_test` and `system_test`. Some sample tests
that fit into the 3 categories are added for examination purposes.

**Benefits:** This will help us to structure our test-suite to better
reflect localised tests from intermodule and interpackage tests.

**Possible Drawbacks:** 3 new markers to work with.

**Related GitHub Issues:**

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>