`Controlled.wires` does not contain work wires (#5728)
**Context:**
`Controlled.wires` contained `work_wires` that are typically used for
the decomposition. This was inconsistent with the general usage of the
attributes. The expected wires behaviour was in the `active_wires`.

**Description of the Change:**
`active_wires` was renamed to `wires`, and `work_wires` has been removed
from the returned wires for `self.wires`

**Benefits:**
- More consistent behaviour of wiring attributes

[[55898](https://app.shortcut.com/xanaduai/story/55898)]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>