Expand logging configuration support in PennyLane, and also target Catalyst (#5528)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** To better unify PennyLane's logging support, updates to
control PennyLane's logging behaviour enabling use with Catalyst allow a
unified way to examine execution pipelines throughout the Pythonic
layer. This PR expands the configuration of logging, allowing direct
editing by a user, as well as adds new module-level log-levels for
Catalyst, and expands jax to also enable logging with jaxlib. The above
will be useful to define interactions across components in the
ecosystem, and aid in building a model of explicit interactions across
the packages.

**Description of the Change:** As above.

**Benefits:** Expands the logging framework to be more configurable, and
extends upcoming support for Catalyst.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>