Added |2> -> |1> amplitude damping to amplitude damping. (#5757)
**Context:**
I recently added `qml.QutritAmplitudeDamping` based on specifications
from this [paper](https://doi.org/10.48550/arXiv.1905.10481). This paper
only considers the transition from |2> to |0>, in reality on a super
conducting device it is more likely to have a |2> to |1> transition.
This PR adds a variable and a new Kraus matrix to
`qml.QutritAmplitudeDamping` to allow for the simulation of the |2> to
|1> transition.

**Description of the Change:**
Adds a new variable, \gamma_3, to `qml.QutritAmplitudeDamping` that
defines the new Kraus matrix K_3. This new Kraus operator allows for the
simulation of |2> to |1> relaxations.

**Benefits:**
Allows for simulation of |2> to |1> relaxation. This will make for more
accurate simulations and is useful for measurement error for qutrits.

**Possible Drawbacks:**
This changes `qml.QutritAmplitudeDamping` and adds an extra parameter,
the naming may be a bit more confusing.

**Related GitHub Issues:**
N/A

---------

Co-authored-by: Olivia Di Matteo <2068515+glassnotes@users.noreply.github.com>