Add separate private qchem functions for external backends (#5553)
**Context:**
The `molecular_hamiltonian` function currently supports three backends:
the in-built differentiable Hartree-Fock backend, the `pyscf` backend,
and the `openfermion` backend. This PR creates dedicated functions for
calling these backends and puts them in proper modules. These changes
are not user-facing but help with easier maintenance and extension of
qchem functions.

**Description of the Change:**
1. Create `_openfermion_hamiltonian` function for building a Hamiltonian
with the openfermion backend.
2. Move the `molecular_hamiltonian` function to `hamiltonian.py.`

**Benefits:**
Easier maintenance and extension of qchem functions.

**Possible Drawbacks:**
None

**Related GitHub Issues:**
None