Bugfix: `TorchLayer` does not work correctly with broadcasting and `tuple` returns (#5816)
**Context:** This bug was caught using the following code:

```
import numpy as np
import pennylane as qml
import torch

n_qubits = 2
dev = qml.device("default.qubit", wires=n_qubits)


@qml.qnode(dev)
def qnode(inputs, weights):
    qml.templates.AngleEmbedding(inputs, wires=range(n_qubits))
    qml.templates.StronglyEntanglingLayers(weights, wires=range(n_qubits))
    return qml.expval(qml.Z(0)), qml.expval(qml.Z(1))


weight_shapes = {"weights": [3, n_qubits, 3]}

qlayer = qml.qnn.TorchLayer(qnode, weight_shapes)
x = torch.tensor(np.random.random((5, 2)))  # Batched inputs with batch dim 5 for 2 qubits


qlayer.forward(x)
```

The `forward` function, in `pennylane/qnn/torch.py`, calls the
`_evaluate_qnode` function under the hood. The issue is caused by the
fact that the latter only works with tuples if these contain lists of
`torch.Tensor`. In the code above, the `_evaluate_qnode` function is
called with a tuple of `torch.Tensor`, which causes the error.

**Description of the Change:** We intercept such a case to temporarily
transform a tuple of `torch.Tensor` into a tuple of lists so that the
code works as originally expected. Obviously, this is not the only
possible solution to the problem. Still, it is a non-invasive way to
solve the problem without changing the original intended tuples workflow
(which is beyond the scope of this PR).

**Benefits:** The above code no longer throws an error.

**Possible Drawbacks:** None that I can think of caused by this PR since
we are not changing the original workflow. We are just fixing an
existing bug by intercepting a specific case that can occur.

**Related GitHub Issues:** #5762 

**Related Shortcut Stories:** [sc-64283]