Implementing Exact Tensor Network method for `default.tensor` (#5786)
**Context:** Currently, the new `default.tensor` device can only
simulate quantum circuits using the Matrix Product State (MPS) method,
based on `quimb`'s `CircuitMPS` class.

**Description of the Change:** With this PR, the device also supports
the Exact Tensor Network method based on `quimb`'s `Circuit` class. This
completes the implementation of the `method` keyword argument, which can
be set to either ``mps`` or ``tn``. For this quarter, this is expected
to be the last feature for the `default.tensor` device (except for
refinements, bug fixes, and performance improvements).

**Benefits:** A new simulation method is implemented for this device,
enabling new functionalities for users.

**Possible Drawbacks:** This PR just implements an additional method,
but it does not modify the existing functionalities (except for some
minor improvements and clarifications in the documentation).

**Related GitHub Issues:** None.

**Related Shortcut Stories:** 
[sc-60805] 
[sc-64011]
[sc-64284]