Fix wire handling on special control wires (#5856)
**Context:**

A user was doing `qml.CRY(0.1, wires=((0,1), 2)`. In this case, the
control wire should be a single wire with a value `(0,1)`. Yes, this is
weird, but this is how wires work in pennylane. Instead, it was being
interpretted as having two control wires, even though `CRY` is only for
singly-controlled `RY` gates.

**Description of the Change:**

Some minor slicing changes in how we handle special controlled ops.  

Now we have:
```
>>> qml.CZ(wires=((0,1), 2))
CZ(wires=[(0, 1), 2])
```

Which is consistent with how we treat wires elsewhere in pennylane.

**Benefits:**

Don't end up with a two qubit gate on three qubits.

**Possible Drawbacks:**

Still weird, and still not what the user was looking for when they
provided wires like that. But at least now they'll get a better error.

**Related GitHub Issues:**