Implement `of.QubitOperator` <-> `qml.ops.LinearCombination`. (#5773)
**Context:**
This PR adds the functionality to convert a ``QubitOperator`` in
OpenFermion to a ``LinearCombination`` in PennyLane and vice versa.

**Description of the Change:**
Added `convert_openfermion` module to `qchem` where users are able to
convert a `QubitOperator` object from OpenFermion into a
`LinearCombination` object in PennyLane using `from_openfermion`.
Conversion of a `LinearCombination` object to `QubitOperator` object is
implemented in `to_openfermion`.

**Benefits:**
Directly convert the operator in terms of a ``LinearCombination`` object
instead of the current functionality where a list of coefficients and
operators is needed (see `_pennylane_to_openfermion` and
`_openfermion_to_pennylane` methods in
[convert.py](https://github.com/PennyLaneAI/pennylane/blob/master/pennylane/qchem/convert.py)).

**Possible Drawbacks:**
If we provide a custom wire mapping where we e.g. swap the order of the
wires, the `LinearCombination` object is still given in the order of the
unswapped wires. However, the order of the operators should only matter
for printing anyway.
```
>>> q_op = openfermion.QubitOperator("X0", 1.2) + openfermion.QubitOperator("Z1", 2.4) + openfermion.QubitOperator("Y2", 0.1)
>>> q_op
1.2 [X0] +
2.4 [Z1] +
0.1 [Y2]
>>> pl_linear_combination = qml.from_openfermion(q_op, wires={0: 2, 1: 1, 2: 0})
>>> pl_linear_combination
1.2 * X(2) + 2.4 * Z(1) + 0.1 * Y(0)
```

**Related GitHub Issues:**
Closes #5759.

---------

Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: soranjh <soran.jahangiri@gmail.com>