Snapshots for arbitrary devices and sample-based measurements (#5805)
**Context:**
`qml.Snapshot` and `qml.snapshots` are debugging tools that allow users
to carry out state measurements at an arbitrary point of the execution.
In #4908, support for arbitrary measurements was added, but limited to
analytical calculations. In addition, support was limited to only some
simulators. Therefore, `qml.snapshots` has to be changed to a transform
that is able to handle any device.

**Description of the Change:**
- For unsupported devices, `qml.snapshots` now splits the `QuantumTape`
into several, creating one for each snapshot measurement, aggregating
all the operations occurring before that snapshot. The result from each
execution is collected into a dictionary as what used to happen before.
- `validate_measurements` is now complemented to take into account the
measurements in the snapshots as part of the validations process.
- Support is also added for sample-based measurements which inherit the
shots from the device.
- Several bugs have been fixed:
- The measurement is now passed to the new snapshot instance during
adjoint and control operations.
- An error is now raised if a tag of type other than string to passed,
as declared in the docs.
- For some devices (e.g. `qutrit.mixed`), the state was returned by the
snapshot regardless of the requested measurement.
- For other devices (e.g. `qubit.legacy`), an error is raised when a
measurement other than `qml.state` is requested, as modifying the code
to be more generic required a lot of effort.
- For legacy devices, the snapshot handling logic did not reset the
`_debugger` attribute after the execution. This is now fixed.
- `default.qutrit` was falsely flagged as snapshot-compatible through
the `_debugger` attribute. This is now commented out.

**Benefits:**
- More consistent snapshot behaviour across devices and clearer error
handling for invalid cases.
- Support for QPUs and majorly non-PL/exotic simulators. 

[[sc-50047](https://app.shortcut.com/xanaduai/story/50047)] Supersedes
#5722

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>