Added repetitions to UCCSD template (#5801)
**Context:**
As discussed in #5741, the UCCSD template should accept an optional
argument `reps` which defines the number of times the UCCSD template is
repeated

**Description of the Change:**
`qml.UCCSD` now accepts an additional optional argument `reps` which is
an integer and defines the number of times the UCCSD template is
repeated. This can improve the accuracy of the UCCSD template by
reducing the Trotter-error but results in deeper circuits. The provided
weights should have the shape `(reps, len(s_wires) + len(d_wires),)`. If
`reps=1`, which is the default value, the weights can also have the
shape `(len(s_wires) + len(d_wires),)`. Therefore, already existing code
using the UCCSD template does not break.

**Benefits:**
The change can improve the accuracy of the UCCSD template by reducing
the Trotter-error.

**Possible Drawbacks:**
 The change results in deeper circuits depending on the value of `reps`.

**Related GitHub Issues:**
#5741

---------

Co-authored-by: Utkarsh <utkarshazad98@gmail.com>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>
Co-authored-by: soranjh <soran.jahangiri@gmail.com>